Added footer; Version upgrade to fix bugs
