--- src/components/Footer/Footer.tsx
@@ -1,5 +1,0 @@
-function Footer() {
-
-}
-
-export default Footer
